goodfaith

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	printf("size of char: %lu byte(s)\n", sizeof(char));
-	printf("size of int: %lu byte(s)\n", sizeof(int));
-	printf("size of long int: %lu byte(s)\n", sizeof(long int));
-	printf("size of long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("size of float: %lu byte(s)\n", sizeof(float));
+	printf("size of char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("size of int: %lu byte(s)\n",(unsigned long)sizeof(int));
+	printf("size of long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("size of long long int: %lu byte(s)\n",(unsigned long)sizeof(long long int));
+	printf("size of float: %lu byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
